Resolved issue of output file being named `.hex`.
snprintf's contract does not allow overlapping buffers, meaning you
cannot use it to append to a given string.  This was failing on Linux
with `gcc.`

In addition, the original code assumed that the extesion was ".ext"
style, or a total of four bytes.  Added a routine that added better
removing of existing extensions if any.